pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-memory-mcp"
version = "0.1.0"
description = "Persistent Project Memory for AI Models"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Danny Maaz" }]
keywords = [
  "Project Memory MCP",
  "AI project memory",
  "Supabase persistent context",
  "AI agent memory",
  "OpenCode memory",
  "Claude Code CLI memory",
  "Qwen Code integration",
  "Codex plugin memory"
]
dependencies = [
  "mcp>=0.1.0",
  "supabase>=2.0.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.0.0"
]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Software Development :: Libraries"
]

[project.scripts]
project-memory-mcp = "src.server:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["src", "src.*", "tests", "tests.*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
